[modify] コマンドラインの引数で実行処理を指定できるよう flag を導入
